Updated Tutorial.odt (declared out-of-date)
The tutorial has not been updated in a while, and while it hasn't been updated, the API has. Therefore, the tutorial is now obsolete and difficult to use. A warning has been issued on the top of the first page of the tutorial. and the actual tutorial will be updated at a later point in time.