Completed the build animation method
The build animation method was completed in heap modus and some modus
images were fixed. An fmi was created for the heap modus as well.
Warning: while all the methods in this version are filled, the complete
modus does not work at the moment. The branch is currently in a
debugging state